Add Cavy tests
