+++ Example/package.json
@@ -16,6 +16,7 @@
   "devDependencies": {
     "babel-jest": "20.0.3",
     "babel-preset-react-native": "2.0.0",
+    "cavy": "^0.5.0",
     "jest": "20.0.4",
     "react-test-renderer": "16.0.0-alpha.12"
   },
