+++ Example/specs/index.js
@@ -1,0 +1,7 @@
+import basicModal from './basicModal';
+import positionTop from './positionTop';
+import positionCentered from './positionCentered';
+import positionBottom from './positionBottom';
+import backdrop from './backdrop';
+
+export default [basicModal, positionTop, positionCentered, positionBottom, backdrop];
